Migrate vlob_create tests
